requirements.txt
datasets
haystack-ai
tqdm
wn
spacy
lemminflect
